+++ Circles/Views/Photo Galleries/GallerySettingsView.swift
@@ -25,8 +25,14 @@ struct GallerySettingsView: View {
     @ViewBuilder
     var generalSection: some View {
         Section("General") {
-            Text("Gallery name")
-                .badge(room.name ?? "(none)")
+            HStack {
+                Text("Gallery name")
+                Spacer()
+                Text(room.name ?? "(none)")
+                    .lineLimit(4)
+                    .multilineTextAlignment(.leading)
+                    .frame(maxWidth: 200)
+            }
             
             HStack {
                 Text("Cover image")