+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add the ability to see posts from a single timeline that you're following
 * Bugfix: Enable deleting other login sessions, but not the current one
 * Bugfix: Display the "Leave Group?" confirmation next to the group to leave
+* Bugfix: Prevent circle/group/gallery names overflowing their visual space
 
 # v1.0.3
 * Bugfix: Ignore invalid config object in account data